State pruning (#216)
* Started work on state db

* Updated for a new hash type

* Pruning and tests

* Generalize on the block hash/key type

* Integrate with the client backend

* Merge w master

* CLI options

* Updated for light client refactoring

* Used IntoIterator

* Fixed invalid input hadling

+++ substrate/runtime/democracy/src/lib.rs
@@ -117,7 +117,7 @@ impl<T: Trait> Module<T> {
 	/// Get the amount locked in support of `proposal`; `None` if proposal isn't a valid proposal
 	/// index.
 	pub fn locked_for(proposal: PropIndex) -> Option<T::Balance> {
-		Self::deposit_of(proposal).map(|(d, l)| d * T::Balance::sa(l.len()))
+		Self::deposit_of(proposal).map(|(d, l)| d * T::Balance::sa(l.len() as u64))
 	}
 
 	/// Return true if `ref_index` is an on-going referendum.
